move actions and reducers to redux dir
