+++ package.json
@@ -18,6 +18,7 @@
     "react-navigation": "^3.11.1",
     "react-redux": "^7.1.0",
     "redux": "^4.0.4",
+    "redux-persist": "^5.10.0",
     "styled-components": "^4.3.2"
   },
   "devDependencies": {
